fix(parity): ratchet Frida replay to exact parity
Unify the latest-only Frida/debug capture contract and ratchet Python and Zig replay behavior to exact parity with the original capture.

+++ src/crimson/bonuses/medikit.py
@@ -1,9 +1,10 @@
 from __future__ import annotations
 
+from ..math_parity import f32
 from .apply_context import BonusApplyCtx
 
 
 def apply_medikit(ctx: BonusApplyCtx) -> None:
     if float(ctx.player.health) >= 100.0:
         return
-    ctx.player.health = min(100.0, float(ctx.player.health) + 10.0)
+    ctx.player.health = min(100.0, float(f32(float(ctx.player.health) + 10.0)))
